app/requirements.txt
numpy==1.10.4
pandas==0.18.0
graphlab-create==1.1
scipy==0.17.0
matplotlib==1.5.1
seaborn==0.7.0
